Move SDK folder
