+++ override/sdk/service/cloudwatchlogs/cloudwatchlogsiface/interface.go
@@ -12,7 +12,7 @@ import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/request"
 
-	"github.com/amazon-contributing/opentelemetry-collector-contrib/sdk/service/cloudwatchlogs"
+	"github.com/amazon-contributing/opentelemetry-collector-contrib/override/sdk/service/cloudwatchlogs"
 )
 
 // CloudWatchLogsAPI provides an interface to enable mocking the
