+++ override/sdk/service/cloudwatchlogs/errors.go
